Closes https://github.com/dotnet/runtime/issues/104954

Currently, unboxing for Nullable<> always go through a helper call, while we can inline it in JIT.
```cs
int? Test(object o) => (int?)o;
```
#### Current codegen:
```asm
; Method Proga:Test(System.Object):System.Nullable`1[int]:this (FullOpts)
G_M11066_IG01:
       sub      rsp, 40
G_M11066_IG02:
       lea      rcx, [rsp+0x20]
       mov      r8, rdx
       mov      rdx, 0x7FF7E261B750      ; System.Nullable`1[int]
       call     CORINFO_HELP_UNBOX_NULLABLE
       mov      rax, qword ptr [rsp+0x20]
G_M11066_IG03:
       add      rsp, 40
       ret      
; Total bytes of code: 37
```
#### New codegen:
```asm
; Method Proga:Test(System.Object):System.Nullable`1[int]:this (FullOpts)
G_M11066_IG01:
       sub      rsp, 40
G_M11066_IG02:
       test     rdx, rdx ;; is null?
       je       SHORT G_M11066_IG04
G_M11066_IG03:
       mov      rax, 0x7FF7E23674D0      ; System.Int32
       cmp      qword ptr [rdx], rax ;; is object's type int32?
       je       SHORT G_M11066_IG05
       jmp      SHORT G_M11066_IG08
G_M11066_IG04:
       xor      edx, edx
       mov      qword ptr [rsp+0x20], rdx
       jmp      SHORT G_M11066_IG06
G_M11066_IG05:
       mov      byte  ptr [rsp+0x20], 1
       mov      eax, dword ptr [rdx+0x08]
       mov      dword ptr [rsp+0x24], eax
G_M11066_IG06:
       mov      rax, qword ptr [rsp+0x20]
G_M11066_IG07:
       add      rsp, 40
       ret      

; cold path
G_M11066_IG08:
       lea      rcx, [rsp+0x20]
       mov      r8, rdx
       mov      rdx, 0x7FF7E263B750      ; System.Nullable`1[int]
       call     CORINFO_HELP_UNBOX_NULLABLE
       jmp      SHORT G_M11066_IG06
; Total bytes of code: 82
```

#### Benchmark:
```cs
using BenchmarkDotNet.Attributes;

public class MyBench
{
    [Benchmark]
    [Arguments(42)]
    [Arguments(null)]
    public int? Unbox(object o) => (int?)o;
}
```
```
BenchmarkDotNet v0.13.12, Ubuntu 22.04.4 LTS (Jammy Jellyfish)
Unknown processor (which is "Ampere Altra")
  Job-MMNLPR : .NET 9.0.0 (42.42.42.42424), Arm64 RyuJIT AdvSIMD
  Job-QXOQTU : .NET 9.0.0 (42.42.42.42424), Arm64 RyuJIT AdvSIMD
```
| Method | Toolchain               | o  | Mean       | Error     | Ratio |
|------- |------------------------ |--- |-----------:|----------:|------:|
| **Unbox**  | **Main** | **?**  |  **7.1195 ns** | **0.0098 ns** |  **1.00** |
| Unbox  | PR | ?  |  0.6698 ns | 0.0072 ns |  0.09 |
|        |                         |    |            |           |       |
| **Unbox**  | **Main** | **42** | **10.9406 ns** | **0.0175 ns** |  **1.00** |
| Unbox  | PR | 42 |  2.7002 ns | 0.0001 ns |  0.25 |