@EgorBo 2024-07-18
@EgorBot -intel -arm64
```cs
using BenchmarkDotNet.Attributes;

public class MyBench
{
    [Benchmark]
    [Arguments(42)]
    [Arguments(null)]
    public int? Unbox(object o) => (int?)o;
}
```