@huoyaoyuan 2024-07-18
Does it worth to recognize `RuntimeHelpers.Unbox_Nullable` and do the same expansion?

Or, can we write the method in C# and make helpers inlinable? This will help more around nullables.