@EgorBo 2024-07-18
@jakobbotsch @AndyAyersMS cc @dotnet/jit-contrib PTAL, the actual change is 40 LOC, benchmarks attached. Jit-diff: https://github.com/MihuBot/runtime-utils/issues/541 obviously, a size regression, but there were also a few cases where jit can now fold the `if` we emit here, e.g. previously we never tried to fold `CORINFO_HELP_UNBOX_NULLABLE(null)` etc.