@EgorBo 2024-07-18
Might be good idea to limit it, yeah

@EgorBot -intel -arm64
```cs
using System.Runtime.CompilerServices;
using BenchmarkDotNet.Attributes;

public class MyBench2
{
    [InlineArray(100)]
    public struct LargeStruct
    {
        public long Fld;
    }

    [Benchmark]
    [Arguments(42)]
    [Arguments(null)]
    public LargeStruct? Unbox(object o) => (MyStruct?)o;
}
```