@EgorBo 2024-07-18
@EgorBot -intel -arm64
```cs
using System.Runtime.CompilerServices;
using BenchmarkDotNet.Attributes;

public class MyBench2
{
    [InlineArray(100)]
    public struct LargeStruct
    {
        public long Fld;
    }

    public IEnumerable<object> TestData()
    {
        yield return null;
        yield return new LargeStruct();
    }

    [Benchmark]
    [ArgumentsSource(nameof(TestData))]
    public LargeStruct? Unbox(object o) => (LargeStruct?)o;
}
```